+++ analyse/src/mitos_halo_spec.c
@@ -224,7 +224,7 @@ int main(int argc, char *argv[]) {
     int *obs_in_bins = calloc(bins, sizeof(int));
 
     /* Transform to momentum space */
-    fftw_complex *fbox = (fftw_complex*) fftw_malloc(N*N*(N/2+1)*sizeof(fftw_complex));
+    fftw_complex *fbox = (fftw_complex*) fftw_malloc((long long)N*N*(N/2+1)*sizeof(fftw_complex));
     fftw_plan r2c = fftw_plan_dft_r2c_3d(N, N, N, box, fbox, FFTW_ESTIMATE);
     fftw_plan c2r = fftw_plan_dft_c2r_3d(N, N, N, fbox, box, FFTW_ESTIMATE);
     fft_execute(r2c);
